@@ -2,17 +2,16 @@
   "name": "lkt/factory-schemas",
   "description": "LKT Factory Schemas",
   "type": "library",
-  "version": "1.0.17",
+  "version": "2.0.0",
   "require": {
-    "php": ">=7.2.0",
-    "lkt/column-types": "^1.0",
+    "php": ">=8.1.0",
     "lkt/value-objects": "^1.0",
     "lkt/tools": "^1.2",
     "lkt/query-builder": "^1.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^7.5.20",
-    "symfony/var-dumper": "^5.0"
+    "phpunit/phpunit": "^9.5",
+    "symfony/var-dumper": "^6.2"
   },
   "license": "MIT",
   "autoload": {
