@@ -4,8 +4,8 @@ namespace Lkt\Factory\Schemas\Traits;
 
 trait FieldWithPaginationOptionTrait
 {
-    protected $itemsPerPage;
-    protected $countableField;
+    protected int $itemsPerPage = -1;
+    protected string $countableField = '';
 
     final public function setPagination(int $itemsPerPage = 20, string $countableField = ''): self
     {
