@@ -6,21 +6,18 @@ use Lkt\Factory\Schemas\Values\BooleanValue;
 
 trait FieldWithSoftTypedOptionTrait
 {
-    protected $softTyped;
+    protected ?BooleanValue $softTyped = null;
 
     /**
      * @param bool $softTyped
      * @return $this
      */
-    final public function setIsSoftTyped(bool $softTyped = true): self
+    final public function setIsSoftTyped(bool $softTyped = true): static
     {
         $this->softTyped = new BooleanValue($softTyped);
         return $this;
     }
 
-    /**
-     * @return bool
-     */
     final public function isSoftTyped(): bool
     {
         if ($this->softTyped instanceof BooleanValue) {
