@@ -6,7 +6,7 @@ use Lkt\Factory\Schemas\Exceptions\InvalidFieldNameException;
 
 final class FieldNameValue
 {
-    private $value;
+    private string $value;
 
     /**
      * @param string $value
@@ -20,9 +20,6 @@ final class FieldNameValue
         $this->value = $value;
     }
 
-    /**
-     * @return string
-     */
     public function getValue(): string
     {
         return $this->value;
