@@ -6,7 +6,7 @@ use Lkt\Factory\Schemas\Fields\AbstractField;
 
 class StringEqualComputedField extends AbstractComputedField
 {
-    final public static function compute(string $name, string $field, string $value = ''): self
+    final public static function compute(string $name, string $field, string $value = ''): static
     {
         $r = new static($name);
         $r->setField($field)->setValue($value);
