@@ -20,7 +20,7 @@ class RelatedField extends AbstractField
         FieldWithMultipleReferencesTrait,
         FieldWithPaginationOptionTrait;
 
-    final public static function defineRelation(string $component, string $name, string $column = ''): self
+    final public static function defineRelation(string $component, string $name, string $column = ''): static
     {
         return (new static($name, $column))->setComponent($component);
     }
