@@ -14,7 +14,7 @@ class RelatedKeysField extends AbstractField
         FieldWithOrderOptionTrait,
         FieldWithSoftTypedOptionTrait;
 
-    final public static function defineRelation(string $component, string $name, string $column = ''): self
+    final public static function defineRelation(string $component, string $name, string $column = ''): static
     {
         return (new static($name, $column))->setComponent($component);
     }
